requirements.txt
# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Azure
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
openai>=1.12.0

# UI
streamlit>=1.31.0
requests>=2.31.0
